chore: frontend UI updates
