--- frontend/src/features/admin-form/create/workflow/components/WorkflowContent/EditStepBlock/components/index.ts
@@ -1,2 +1,0 @@
-export * from './RespondentBlock'
-export * from './SaveActionGroup'
