--- frontend/src/features/admin-form/create/workflow/components/WorkflowContent/InactiveStepBlock/EmailBadge.tsx
@@ -1,18 +1,0 @@
-import { Flex } from '@chakra-ui/react'
-
-import Badge, { BadgeProps } from '~components/Badge'
-
-/**
- * Styled badge for displaying emails
- */
-export const EmailBadge = (props: BadgeProps) => (
-  <Flex>
-    <Badge
-      maxW="100%"
-      variant="subtle"
-      colorScheme="secondary"
-      whiteSpace="pre-wrap"
-      {...props}
-    />
-  </Flex>
-)
